refactoring

+++ include/vb/mhp.h
@@ -18,23 +18,23 @@ extern "C" {
 
 typedef struct vb_mhp_chunk_t {
 	void *ref;
-	size_t size_total;
+	int64_t size_total;
 } vb_mhp_chunk_t;
 
 typedef struct vb_mhp_block_t {
 	struct vb_mhp_block_t *next_block;
 	vb_mhp_chunk_t *first_free_chunk;
-	size_t size_used;
-	size_t size_total;
+	int64_t size_used;
+	int64_t size_total;
 } vb_mhp_block_t;
 
 typedef struct {
 	vb_mhp_block_t *block;
-	size_t size_used;
-	size_t size_total;
-	size_t size_overhead;
-	size_t size_block_init;
-	size_t size_total_max;
+	int64_t size_used;
+	int64_t size_total;
+	int64_t size_overhead;
+	int64_t size_block_init;
+	int64_t size_total_max;
 } vb_mhp_head_t;
 
 typedef struct {
